refactor: create IntegrationPlatformsGrid folder
